# frontend/src/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: TinBoker Backend API
  description: Backend API for TinBoker - Financial podcast insights platform
  version: 1.0.0
servers:
- url: https://api.tinboker.com
  description: Production server
- url: https://staging-api.tinboker.com
  description: Staging server
- url: https://dev-api.tinboker.com
  description: Development server
- url: http://localhost:8000
  description: Local development
paths:
...
  /api/news:
    get:
      tags:
      - news
      summary: Get Sorted News
      description: 'Get sorted news list


        Query params:

        - sort_by: Sort field (date, created_at, updated_at, title)


        CDN Cache: 10 minutes'
      operationId: get_sorted_news_api_news_get
      parameters:
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          description: Sort field
          default: date
          title: Sort By
        description: Sort field
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockEvent'
                title: Response Get Sorted News Api News Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/news/{news_id}:
    get:
      tags:
      - news
      summary: Get News By Id
      description: 'Get news by ID


        Returns complete news event with related tickers


        CDN Cache: 10 minutes'
      operationId: get_news_by_id_api_news__news_id__get
      parameters:
      - name: news_id
        in: path
        required: true
        schema:
          type: string
          description: News ID
          title: News Id
        description: News ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockEvent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StockEvent:
      properties:
        id:
          type: string
          title: Id
          description: Unique event identifier
        type:
          type: string
          enum:
          - earnings
          - conference
          - news
          - dividend
          title: Type
          description: Event type
        date:
          type: integer
          title: Date
          description: Event date (Unix timestamp in milliseconds)
        title:
          type: string
          title: Title
          description: Event title
        description:
          type: string
          title: Description
          description: Event description
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
          description: Event content
        relatedTickers:
          items:
            type: string
          type: array
          title: Relatedtickers
          description: List of related stock tickers
      type: object
      required:
      - id
      - type
      - date
      - title
      - description
      - relatedTickers
      title: StockEvent
      description: Stock-related event (earnings, news, etc.)
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer